+++ src/test/java/com/example/petbuddybackend/service/photo/PhotoServiceTest.java
@@ -18,6 +18,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -25,6 +26,8 @@ import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.time.LocalDateTime;
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -85,9 +88,28 @@ public class PhotoServiceTest {
         storageClientMock.close();
     }
 
+    @Test
+    void findPhotoLinkByNullableId_nullBlob_shouldReturnEmpty() {
+        Optional<PhotoLink> result = firebasePhotoService.findPhotoLinkByNullableId(null);
+
+        assertTrue(result.isEmpty());
+        verify(photoRepository, never()).findById(any());
+    }
 
     @Test
-    public void testUploadPhoto_validFile_shouldUploadSuccessfully() throws IOException {
+    void findPhotoLinkByNullableId_blobNotNull_shouldReturnPhotoLink() {
+        PhotoLink photo = new PhotoLink(BLOB_PATH, PHOTO_URL, LocalDateTime.now().plusDays(10));
+        when(photoRepository.findById(BLOB_PATH)).thenReturn(Optional.of(photo));
+
+        Optional<PhotoLink> result = firebasePhotoService.findPhotoLinkByNullableId(BLOB_PATH);
+
+        assertTrue(result.isPresent());
+        assertEquals(photo, result.get());
+        verify(photoRepository, times(1)).findById(BLOB_PATH);
+    }
+
+    @Test
+    void testUploadPhoto_validFile_shouldUploadSuccessfully() throws IOException {
         when(mockBucket.create(any(String.class), any(ByteArrayInputStream.class), any(String.class))).thenReturn(mockBlob);
         when(mockBlob.signUrl(anyLong(), any())).thenReturn(new URL("http://signedurl.com"));
         when(tika.detect(any(InputStream.class))).thenReturn("image/jpeg");
@@ -102,7 +124,7 @@ public class PhotoServiceTest {
     }
 
     @Test
-    public void testUploadPhoto_invalidFile_shouldThrowInvalidPhotoException() throws IOException {
+    void testUploadPhoto_invalidFile_shouldThrowInvalidPhotoException() throws IOException {
         MockMultipartFile invalidPhoto = new MockMultipartFile("file", "", "text/plain", new byte[]{0});
         when(tika.detect(any(InputStream.class))).thenReturn("text/plain");
 
@@ -112,7 +134,53 @@ public class PhotoServiceTest {
     }
 
     @Test
-    public void testDeletePhoto_validBlob_shouldDeleteSuccessfully() {
+    void uploadPhotos_emptyListPassed_shouldReturnEmptyList() {
+        List<PhotoLink> returnedList = firebasePhotoService.uploadPhotos(Collections.emptyList());
+
+        assertTrue(returnedList.isEmpty());
+        verify(photoRepository, never()).saveAll(anyList());
+    }
+
+    @Test
+    void uploadPhotos_allPhotosUploadedSuccessfully() throws IOException {
+        List<MultipartFile> inputList = List.of(validPhoto, validPhoto);
+
+        when(mockBucket.create(any(String.class), any(ByteArrayInputStream.class), any(String.class))).thenReturn(mockBlob);
+        when(mockBlob.signUrl(anyLong(), any())).thenReturn(new URL("http://signedurl.com"));
+        when(tika.detect(any(InputStream.class))).thenReturn("image/jpeg");
+        when(photoRepository.save(any(PhotoLink.class))).thenAnswer(invocation -> invocation.getArgument(0));
+
+        firebasePhotoService.uploadPhotos(inputList);
+
+        verify(photoRepository, times(1)).saveAll(anyList());
+    }
+
+    @Test
+    void uploadPhotos_secondPhotoUploadFails_shouldRollback() throws IOException {
+        MultipartFile invalidPhoto = mock(MultipartFile.class);
+        List<MultipartFile> inputList = List.of(validPhoto, invalidPhoto);
+
+        // Return valid on first invoke
+        when(mockBucket.create(any(String.class), any(InputStream.class), eq(validPhoto.getContentType()))).thenReturn(mockBlob);
+
+        // Throw on second
+        when(mockBucket.create(any(String.class), any(InputStream.class), any(String.class))).thenThrow(RuntimeException.class);
+
+        when(mockBlob.signUrl(anyLong(), any())).thenReturn(new URL("http://signedurl.com"));
+        when(invalidPhoto.getInputStream()).thenReturn(new ByteArrayInputStream(new byte[]{0}));
+        when(tika.detect(any(InputStream.class))).thenReturn("image/jpeg");
+        when(photoRepository.save(any(PhotoLink.class))).thenAnswer(invocation -> invocation.getArgument(0));
+        when(invalidPhoto.getOriginalFilename()).thenReturn("invalid.jpg");
+
+        assertThrows(RuntimeException.class,
+                () -> firebasePhotoService.uploadPhotos(inputList));
+
+        verify(photoRepository, never()).saveAll(anyList());
+        verify(photoRepository, times(1)).deleteAll(anyList());
+    }
+
+    @Test
+    void testDeletePhoto_validBlob_shouldDeleteSuccessfully() {
         PhotoLink photo = new PhotoLink(BLOB_PATH, PHOTO_URL, LocalDateTime.now().plusDays(10));
         when(photoRepository.findById(BLOB_PATH)).thenReturn(Optional.of(photo));
         when(mockBucket.get(BLOB_PATH)).thenReturn(mockBlob);
@@ -124,7 +192,7 @@ public class PhotoServiceTest {
     }
 
     @Test
-    public void testDeletePhoto_nonExistentBlob_shouldNotThrowException() {
+    void testDeletePhoto_nonExistentBlob_shouldNotThrowException() {
         String nonExistentBlob = "non/existent/blob";
 
         when(photoRepository.findById(nonExistentBlob)).thenReturn(Optional.empty());
@@ -133,7 +201,18 @@ public class PhotoServiceTest {
     }
 
     @Test
-    public void testUpdatePhotoExpiration_withinThreshold_shouldNotRenew() {
+    void testDeletePhotos_shouldRemoveAllPhotos() {
+        List<PhotoLink> photos = List.of(
+                new PhotoLink("valid/blob/path", "http://signedurl.com", LocalDateTime.now()),
+                new PhotoLink("valid/blob/path", "http://signedurl.com", LocalDateTime.now())
+        );
+
+        firebasePhotoService.deletePhotos(photos);
+        verify(photoRepository, times(2)).delete(any());
+    }
+
+    @Test
+    void testUpdatePhotoExpiration_withinThreshold_shouldNotRenew() {
         LocalDateTime expiresAt = LocalDateTime.now().plusDays(20);
         PhotoLink photo = new PhotoLink("valid/blob/path", "http://signedurl.com", expiresAt);
 
@@ -142,7 +221,7 @@ public class PhotoServiceTest {
     }
 
     @Test
-    public void testUpdatePhotoExpiration_expired_shouldRenew() throws MalformedURLException {
+    void testUpdatePhotoExpiration_expired_shouldRenew() throws MalformedURLException {
         LocalDateTime expiresAt = LocalDateTime.now().minusSeconds(1);
         PhotoLink photo = new PhotoLink("valid/blob/path", "http://signedurl.com", expiresAt);
         when(mockBucket.get(photo.getBlob())).thenReturn(mockBlob);
@@ -151,5 +230,22 @@ public class PhotoServiceTest {
         firebasePhotoService.updatePhotoExpiration(photo);
 
         verify(photoRepository, times(1)).save(photo);
+    }
+
+    @Test
+    void testUpdatePhotoExpirations_shouldRenewAllExpiredPhotos() throws MalformedURLException {
+        LocalDateTime expiredDate = LocalDateTime.now().minusSeconds(1);
+        LocalDateTime validDate = LocalDateTime.now().plusDays(10);
+        PhotoLink expiredPhoto = new PhotoLink("valid/blob/path", "http://signedurl.com", expiredDate);
+        PhotoLink validPhoto = new PhotoLink("valid/blob/path", "http://signedurl.com", validDate);
+        when(mockBucket.get(expiredPhoto.getBlob())).thenReturn(mockBlob);
+        when(mockBlob.signUrl(anyLong(), any())).thenReturn(new URL("http://new-signedurl.com"));
+
+        List<PhotoLink> photos = List.of(expiredPhoto, validPhoto);
+        firebasePhotoService.updatePhotoExpirations(photos);
+
+        verify(photoRepository, times(1)).saveAll(photos);
+        assertNotEquals(expiredDate, expiredPhoto.getUrlExpiresAt());
+        assertEquals(validDate, validPhoto.getUrlExpiresAt());
     }
 }
